Remove context.interBuffer
This code is not actually used for anything right now (no tests fail).
The code that was attempting to use this had a bug in it so that
interBuffer would never get populated.

This code seems to be trying to handle the dcs intermediate state.
Perhaps we will want to support this state correctly, but since it
does not currently work we're going to remove it for now to simplify
our refactoring.