Refactor to emit events to channel
In order to not have to rewrite all the tests there is currently an
adapter that will call the callback interface.

Because this is now async the tests now have to wait for the event to
actually come through. This will probably make the tests flaky in an
environment like github actions.