Make the generic event private
We specifically do not want consumers casting these events to a
concrete type. Instead we want the consumer to just call Raw() and
deal with that data.

That allows for safe additions to the parser without breaking existing
consumers. The consumer does not have to handle every type as long as
they have a default case that handles the event via Raw() their code
will continue to work after upgrading this package.